chore(tool): consolidate setup package tests into per-source-file pairing
Fifth package in the test-consolidation pass for #1128. tool/internal/setup
had five orphaned test files, each a grab-bag spanning multiple source
files rather than a single-target merge:

  extra_helpers_test.go splits two ways:
    TestStateManagerDiscard, TestStateManagerDiscardEmpty -> state_test.go
    TestGenerateRuntimePerPackageSkipsPackagesWithoutFiles -> setup_test.go

  logger_delegate_test.go:
    TestSetupPhaseLogDelegators -> setup_test.go (SetupPhase's log
    delegator methods are declared in setup.go)

  match_one_rule_test.go:
    TestMatchOneRule (+ its matchOneRuleSource/parseMatchSource
    fixtures) -> match_test.go (matchOneRule is in match.go)

  pin_helpers_test.go splits three ways:
    TestIsSetup                        -> setup_test.go (isSetup)
    TestPrepareVendoredBuild (+newModuleDir) -> pin_test.go (prepareVendoredBuild)
    TestFindGoSources                  -> find_test.go (findGoSources)

  setup_extra_test.go splits two ways:
    13 tests targeting createRuleFromFields, parseRuleFromYaml, runMatch,
    preciseMatching, rulesFromDir, loadCustomRules, matchDeps, and
    loadRules -> match_test.go (all declared in match.go)
    3 tests targeting findToolFiles, collectImports, and
    walkInstrumentation -> config_test.go (all declared in config.go)

lock.go/lock_windows.go/lock_nonwindows.go were already correctly
paired with their own _test.go files and are untouched.

This is a pure move: no test was added, removed, renamed, or edited.
The {name, pass/fail} set is identical before and after the merge.
